requirements.txt
yfinance
pandas
apache-airflow
